transgov/src/requirements.txt
dj-database-url==0.4.2
Django==1.11.8
django-crispy-forms==1.7.0
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-haystack==2.5.1
django-jquery==3.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
elasticsearch==5.0.1
gunicorn==19.7.1
olefile==0.44
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
PyJWT==1.5.3
sqlparse==0.2.4
urllib3==1.22
